requirements.txt
# Core app
streamlit
python-dotenv

# LangChain
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-groq

# Vector DB
chromadb

# Embeddings
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
# if problem pip install --force-reinstall torch --index-url https://download.pytorch.org/whl/cpu
# python -c "import torch, numpy; print(torch.__version__, numpy.__version__, torch.cuda.is_available())"



# PDF handling
pypdf

numpy>=1.24.0,<2.0
# if problem pip install "numpy>=1.24,<2.0" --force-reinstall
# check vrsn numpy python -c "import numpy; print(numpy.__version__)"


# remove cache pip cache purge
# pip install -r requirements.txt --force-reinstall

